[lgdo] rename waveform_table waveformtable for consistency
