+++ src/pygama/lgdo/waveformtable.py
